clients/python/requirements.txt
"""RATS Python Client requirements"""

paho-mqtt==1.7.0
